Falta requerimiento opcional - envio de correo masivo al agregar un gasto

+++ gastos.js
@@ -6,10 +6,17 @@ const gastos_arr = gastosJSON.gastos;
 
 //peticion: POST gastos
 const gastosRoommate = (g) => {
-    gastos_arr.push(g);
-    const gastos = gastosRoommates(gastosJSON);
-    fs.writeFileSync('gastos.json', JSON.stringify(gastos));
+    const roommatesJSON = JSON.parse(fs.readFileSync('roommates.json', 'UTF8'));
+    const roommates_arr = roommatesJSON.roommates;
+    if (roommates_arr.length > 0) { //Se valida que existan roommates para no crear gastos null
+        gastos_arr.push(g);
+        const gastos = gastosRoommates(gastosJSON);
+        fs.writeFileSync('gastos.json', JSON.stringify(gastos));
+    }else{
+        console.log('No se han agregado roommates');
+    }
 }
+
 //peticion: DELETE gastos
 const deleteGasto = (id) => {
     gastosJSON.gastos = gastos_arr.filter((g) => g.id !== id);
